Added basic support for storing data in HttpContext.Session across one redirect. Currently supports any object, so all customization is up to the user.
Had to make ActionHttpHandler implement IRequiresSessionState in order to make Asp.Net kick start the SessionStateModule stuff.

Added IRequestDataProvider and RequestDataProvider as the gateway into and out of getting and setting a single item to/from the session.

Note: I am also using HttpContextBase (which allows for testing) so I setup FrameworkServicesRegistry to build this type for us from HttpContext.Current.